requirements/base.txt
Django==1.6.1
South==0.8.4
django-extensions==1.2.5
ipython==1.1.0
psycopg2==2.5.1
six==1.4.1
wsgiref==0.1.2
pytz==2013.8
pytest==2.5.1
pytest-django==2.4
model-mommy==1.2
pyephem==3.7.5.2